refactor: create claims.rs
